Add script that downloads templates
A script has been added to download the templates from GitHub. The
templates are checked out as a Git repository, making it really easy to
download changes later on.

If the repository does not exist yet, "git clone" is executed to create
a local copy of the it. If it exists, "git pull" is invoked to get the
latest version of the templates.