Remove Python workflow
The Python workflow was clunky and used three different scripts to
achieve its results. Over the years, I honestly forgot how these files
worked together, and found it harder and harder to fix bugs.

The workflow will be reimplemented using Rust. The main reason is that I
want to, and the other two reasons are a chance to write more tests and
to use better dependency management.