Write template names into .gitignore files
The names of the templates that are used to generate a .gitignore file
are now being written into the file as well. Before pasting the contents
of each template, the template's name is written into the file as a
comment. This makes it possible to see which lines were copied from
which template.