Fix and version build cache on GitHub Actions
The cache on GitHub Actions was not working properly, because the hash
for the cache was calculated before the code was checked out. This
resulted in an empty hash, and a cache hit on a very old version of the
dependencies. The code is checked out as the first step now, and the
cache has been versioned so that it is easier to invalidate a broken
state.