Switch to Flowcrafter-managed workflows
FlowCrafter[^1] is a command-line tool that makes it easy to create and
update GitHub Actions workflows from a shared repository. It has been
set up to pull from my personal collection, which is very similar to
what was already configured in this repository. The most meaningful
change is that Rust builds now run inside a pre-built container.